+++ src/main/java/lotto/view/OutputView.java
@@ -14,7 +14,7 @@ public class OutputView {
         System.out.println(rank.toString());
     }
 
-    public static void printRateOfReturn(double rateOfReturn) {
-        System.out.println("총 수익률은 " + (int) rateOfReturn + "%입니다.");
+    public static void printRateOfReturn(int rateOfReturn) {
+        System.out.println("총 수익률은 " + rateOfReturn + "%입니다.");
     }
 }
